WINDOWS x64: Small code changes to remove compiler warnings
Almost all are int/size_t
Need to watch for 'long' vs 'long long'
Hiding the strlen() int/size_t problems for the moment
Changed some pointer to long conversions

We have bigger problems with 64-bit filesizes, seeking etc...

+++ EX_browser_net.c
@@ -559,7 +559,7 @@ int GetServerPingsAndInfosProc(void * lpParameter)
 	extern cvar_t sb_listcache;
 	extern void SB_Serverlist_Serialize_f(void);
 
-	int full = (int) lpParameter;
+	int full = (int)(uintptr_t)lpParameter;
     abort_ping = 0;
 
 	if (full || serversn_passed == 0) {
