+++ hash.c
@@ -128,7 +128,7 @@ void *Hash_GetInsensitive(hashtable_t *table, const char *name)
 }
 void *Hash_GetKey(hashtable_t *table, char *key)
 {
-	int bucknum = ((long) key) % table->numbuckets;
+	int bucknum = ((uintptr_t) key) % table->numbuckets;
 	bucket_t *buck;
 
 	buck = table->bucket[bucknum];
@@ -234,7 +234,7 @@ void *Hash_AddInsensitive(hashtable_t *table, char *name, void *data)
 }
 void *Hash_AddKey(hashtable_t *table, char *key, void *data, bucket_t *buck)
 {
-	int bucknum = ((long) key) % table->numbuckets;
+	int bucknum = ((uintptr_t) key) % table->numbuckets;
 
 	buck->data = data;
 	buck->keystring = key;
@@ -311,7 +311,7 @@ void Hash_RemoveData(hashtable_t *table, char *name, void *data)
 
 void Hash_RemoveKey(hashtable_t *table, char *key)
 {
-	int bucknum = ((long) key) % table->numbuckets;
+	int bucknum = ((uintptr_t) key) % table->numbuckets;
 	bucket_t *buck;	
 
 	buck = table->bucket[bucknum];
